% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_jitter2D.R
\name{compute_jitter2D}
\alias{compute_jitter2D}
\title{Combine 3 parameters for describe jitter in 2D, each parameter has two posible levels.}
\usage{
compute_jitter2D(
  data,
  noise = "quasi",
  dir = "global",
  weight = NULL,
  bw = 0.5,
  seed = NA
)
}
\arguments{
\item{noise:}{Random or Quasirandom}

\item{dir:}{Global or Local}

\item{Voronoi:}{True or False (not implemented yet)}

\item{weight:}{spread factor (copied from ggplot2::geom_jitter, not used)}

\item{bw:}{bandwidth for kernel in local correlation}

\item{seed:}{simulation seed (copied from ggplot2::geom_jitter, not used)}
}
\description{
Combine 3 parameters for describe jitter in 2D, each parameter has two posible levels.
}
